Clean up project and add python 2 support
Vaguely worded commit message for the win.

"Clean up project" in this case means a substantial change to what we have to implement. Instead of lots of different add functions, students should just help implement the `add`, `subtract`, `multiply`, and `delete`. This lets the project have similar instructions for both languages.

The Python 2 implementation includes tests run with `pytest`, and installs the module as an editable module.

This also uses `ava.js` as a test runner for the Node folder. Maybe too much overhead? But it is installed with `npm install`, and there are instructinos in the README.